+++ device/esp_tinyusb/CHANGELOG.md
@@ -11,6 +11,7 @@
 - esp_tinyusb: Removed configuration option to handle TinyUSB events outside of this driver
 - MSC: Removed dedicated callbacks; introduced a single callback with an event ID for each storage
 - MSC: Added storage format support
+- MSC: Added dual storage support (SPI/Flash and SD/MMC)
 - MSC: Updated public API; refer to the [MSC Class Migration guide v2](/docs/device/migration-guides/v2/tinyusb_msc.md)
 - Console: Updated public API; refer to the [Console Class Migration guide v2](/docs/device/migration-guides/v2/tinyusb_console.md)
 - CDC-ACM: Updated public API; refer to the [CDC-ACM Class Migration guide v2](/docs/device/migration-guides/v2/tinyusb_cdc_acm.md)